+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/minerva/MinervaCourseViewHolder.java
@@ -17,7 +17,7 @@ import be.ugent.zeus.hydra.ui.minerva.overview.CourseActivity;
 /**
  * @author Niko Strijbol
  */
-class MinervaCourseViewHolder extends DataViewHolder<Pair<Course, Integer>> implements SearchStateListener {
+class MinervaCourseViewHolder extends DataViewHolder<Pair<Course, Long>> implements SearchStateListener {
 
     private final TextView name;
     private final TextView subtitle;
@@ -50,7 +50,7 @@ class MinervaCourseViewHolder extends DataViewHolder<Pair<Course, Integer>> impl
      * @param data The data.
      */
     @Override
-    public void populate(final Pair<Course, Integer> data) {
+    public void populate(final Pair<Course, Long> data) {
         Course course = data.first;
         name.setText(course.getTitle());
         final CharSequence tutor = Utils.fromHtml(course.getTutorName());
